+++ source/extensions/isaacsim.asset.importer.mjcf/python/impl/converter.py
@@ -22,7 +22,6 @@ import importlib
 import os
 import shutil
 
-import omni
 from isaacsim.asset.importer.utils.impl import (
     importer_utils,
     merge_mesh_utils,
@@ -152,13 +151,9 @@ class MJCFImporter:
         self.stage = None
         gc.collect()
 
-        ext_manager = omni.kit.app.get_app().get_extension_manager()
-        ext_id = ext_manager.get_enabled_extension_id("isaacsim.asset.transformer.rules")
-        extension_path = ext_manager.get_extension_path(ext_id)
+        from isaacsim.asset.transformer.rules import DEFAULT_PROFILE_PATH
 
-        asset_structure_profile_json_path = os.path.normpath(
-            os.path.abspath(os.path.join(f"{extension_path}", "data", "isaacsim_structure.json"))
-        )
+        asset_structure_profile_json_path = DEFAULT_PROFILE_PATH
 
         if self.config.debug_mode:
             log_path = os.path.normpath(